+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.2.0
 
+- Added ability to log whether you joined voice chat in a match [#44](https://github.com/cheshire137/competiwatch-desktop/issues/44)
+- Added support for a 'joined voice' field when importing from a CSV file, and exporting matches, to specify whether you joined voice chat in a match
 - Added Ashe [#48](https://github.com/cheshire137/competiwatch-desktop/issues/48)
 - Added competitive season 13 [#47](https://github.com/cheshire137/competiwatch-desktop/issues/47)